функция validate_parameter_int перенесена в /utils/common.py
