+++ drevo/views/knowledge_grade_view.py
@@ -9,28 +9,7 @@ from drevo.models.relation import Relation
 from drevo.models.relation_grade import RelationGrade
 from drevo.models.relation_grade_scale import RelationGradeScale
 from drevo.utils.knowledge_grader import KnowledgeGraderService
-
-
-def validate_parameter_int(param, default: int = 0, good_values: list = None):
-    """ Валидация параметра param, который может быть числом, строкой или None
-        Если передан список good_values - проверяется на вхождение в него
-        Если параметр равен None - возвращается default
-        Если параметр строка не является числом - возвращается default
-        Если параметр не входит в список good_values - возвращается default
-    """
-    if param is None:
-        return default
-
-    if isinstance(param, str):
-        if param.isnumeric():
-            param = int(param)
-        else:
-            return default
-
-    if good_values and param not in good_values:
-        return default
-
-    return param
+from drevo.utils.common import validate_parameter_int
 
 
 class KnowledgeFormView(LoginRequiredMixin, TemplateView):
